requirements.txt
glfw==2.5.0
numpy==1.22.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
